ublox: define port type hints for the TOBY-L4
The TOBY-L4 exposes multiple TTY ports, but each of them has different
purposes:

The TTY at interface #2 is a fully capable AT port but shouldn't be
used to start connections, so we flag it as secondary.

The TTY at interface #4 is a non-AT debug port, so we explicitly
ignore it to make probing faster.

The TTY at interface #6 is the fully capable AT port that must be used
as primary port.

The TTY at interface #8 is an AT-capable port but dedicated to FOTA
update operations.