broadband-bearer: fix object types in connect_3gpp() steps
Some of the steps in the connect_3gpp() sequence are implemented as
methods in the MMBroadbandBearer class, and therefore the objects
received in the corresponding ready() methods are all bearers, not
modems.

This issue wasn't found before because the objects received were not
really being used after all, as we were relying on the ones kept in
the context.

(cherry picked from commit 1e318bec54da3f37901eadb2a04a4520737eb9a0)