build: explicitly disable incompatible pointer warnings in MM 1.6.x
ModemManager relied a lot on e.g. just implicitly casting the return
of g_object_new() to a more specific type name, so the new checks
added in GObject end up trigger a lot of errors when building with
-Wincompatible-pointer-types.

We just disable those warnings by default for MM 1.6.x builds.

MM 1.8 or git master are NOT affected by this problem.