iface-modem-voice: fix ordering of D-Bus signals for call addition/deletion
To have proper ordering in the D-Bus signals, the skeleton's property
changes must be flushed before the Call{Add,Delet}ed signals are
emitted.  Without this flush, the emission of the PropertiesChanged
signal is delayed until the main loop is idle.  This causes problems
on the client side, for example the CallAdded signal being received
before the Calls property contains the call.

Closes: #81