iface-modem-voice: handle in-call state at modem level
We cannot handle in-call state at call level because the state may
apply to multiple active calls at the same time:
 * the modem is in-call if there is at least one ongoing call (>=1).
 * the modem is not in-call if there are no ongoing calls (==0).

The new logic implements call state monitoring for all available call
objects, and whenever the state of one or more calls change, we'll go
over all of them to see how many of the calls can be considered to be
ongoing (i.e. not terminated, not unknown). If we have at least one
call ongoing, we'll setup the in-call state and otherwise, we'll clean
it up.

Setting up and cleaning up the in-call state involves unsolicited
message and audio channel settings management operations.