mm-broadband-bearer: Increase CGACT timeout on deactivation
On u-blox modems it has been observed that attempting to force down the
bearer when the modem had an active data connection, lost the connection
but the bearer is still up, and the modem is searching for the network,
ModemManager would timeout before the CGACT command could return.  This
can put the modem in a state where the ModemManager state does not match
the modem's actual state of being connected.

When deactivating the current active bearer, the timeout has been
changed from 10 seconds to 45 seconds.  u-blox modems can take up to 40
seconds to respond to deactivation requests through the AT+CGACT
command according to the u-blox AT command reference.