iface-modem: report last attempt to load_unlock_required implementations
The load_unlock_required() step will be automatically retried by the
generic interface logic unless some specific errors happen (e.g. it
won't be retried if a SIM missing error is received).

This patch allows telling the implementation of the method whether
this time being run is the last one before reporting a hard error or
not.