libmm-glib,location: fix 'unknown' lat/long/alt numeric values
G_MINDOUBLE is 0.0, which is a valid altitude, latitude, and
longitude. Change the 'unknown' location values to -G_MAXDOUBLE
instead, which is a completely invalid value in all cases.

This is strictly an ABI break, but it's also true that until now the
methods returning these 'unknown' values could also return valid 0.0
values that would have been detected as 'unknown' by the caller.
So, just stop this madness and don't consider this any ABI break.

This is the same thing we had to do with MM_SIGNAL_UNKNOWN in
fe66bdf65e57fa7dee3dcb8dea068fb3fc7aec34.