bearer-qmi: close QMI port if it was open during connection
If the bearer is using a QMI port that is not the primary one, we need
to explicitly open it during the connection attempt. Keep track of
that, and make sure we close it during normal disconnection or if the
connection attempt fails.

(cherry picked from commit 2790074c7e00b94f5079ba7bda8911a12ab0e28c)